Finish relocating files
