+++ conda/meta.yaml
@@ -12,7 +12,7 @@ build:
   number: 0
   noarch: python
   entry_points:
-    - sdgym=sdgym.__main__:main
+    - sdgym=sdgym.cli.__main__:main
   script: "{{ PYTHON }} -m pip install . -vv"
 
 requirements:
