+++ sdgym/cli/summary.py
@@ -5,7 +5,8 @@ import re
 import numpy as np
 import pandas as pd
 
-from sdgym.s3 import read_csv, write_file
+from sdgym.cli.utils import read_csv
+from sdgym.s3 import write_file
 
 KNOWN_ERRORS = (
     ('Synthesizer Timeout', 'timeout'),
